fixed bug with nn kernel function, finished pacoh_map_gp, apart from some interface changes left to do

+++ pacoh/modules/common.py
@@ -11,6 +11,6 @@ class PositiveParameter(hk.Module):
         self.shape = shape
 
     def __call__(self):
-        exponentiated = jnp.exp(hk.get_parameter("__positive_log_scale_param", shape=self.shape, dtype=jnp.float64,
+        exponentiated = jnp.exp(hk.get_parameter("__positive_log_scale_param", shape=self.shape, dtype=jnp.float32,
                                        init=hk.initializers.Constant(self.raw_init)))
         return exponentiated + self.boundary_value
